# api/audit/v1/openapi.json
{
  "components": {
    "responses": {
      "ProblemResponse": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "Normalized RFC 9457-style Audit problem."
      }
    },
    "schemas": {
...
      "ChainVerification": {
        "additionalProperties": false,
        "allOf": [
          {
            "if": {
              "properties": {
                "complete": {
                  "const": true
                }
              },
              "required": [
                "complete"
              ]
            },
            "then": {
              "properties": {
                "nextSequence": false
              }
            }
          },
          {
            "if": {
              "properties": {
                "complete": {
                  "const": false
                }
              },
              "required": [
                "complete"
              ]
            },
            "then": {
              "required": [
                "nextSequence"
              ]
            }
          }
        ],
        "oneOf": [
          {
            "properties": {
              "installationId": false
            },
            "required": [
              "tenantId"
            ]
          },
          {
            "properties": {
              "tenantId": false
            },
            "required": [
              "installationId"
            ]
          }
        ],
        "properties": {
          "apiVersion": {
            "const": "audit.matrix.xiak.com/v1"
          },
          "complete": {
            "type": "boolean"
          },
          "firstPreviousHash": {
            "$ref": "#/components/schemas/Digest"
          },
          "fromSequence": {
            "maximum": 9007199254740991,
            "minimum": 1,
            "type": "integer"
          },
          "installationId": {
            "$ref": "#/components/schemas/ID"
          },
          "kind": {
            "const": "ChainVerification"
          },
          "lastRecordHash": {
            "$ref": "#/components/schemas/Digest"
          },
          "nextSequence": {
            "maximum": 9007199254740991,
            "minimum": 1,
            "type": "integer"
          },
          "recordCount": {
            "maximum": 10000,
            "minimum": 1,
            "type": "integer"
          },
          "state": {
            "$ref": "#/components/schemas/VerificationState"
          },
          "tenantId": {
            "$ref": "#/components/schemas/TenantID"
          },
          "toSequence": {
            "maximum": 9007199254740991,
            "minimum": 1,
            "type": "integer"
          },
          "verifiedAt": {
            "$ref": "#/components/schemas/Timestamp"
          }
        },
        "required": [
          "apiVersion",
          "complete",
          "firstPreviousHash",
          "fromSequence",
          "kind",
          "lastRecordHash",
          "recordCount",
          "state",
          "toSequence",
          "verifiedAt"
        ],
        "type": "object"
      },
...
      "Digest": {
        "pattern": "^sha256:[0-9a-f]{64}$",
        "type": "string"
      },
...
      "ID": {
        "maxLength": 128,
        "minLength": 1,
        "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$",
        "type": "string"
      },
      "IngestionOutcome": {
        "enum": [
          "ACCEPTED",
          "DUPLICATE"
        ],
        "type": "string"
      },
      "IngestionResult": {
        "additionalProperties": false,
        "properties": {
          "apiVersion": {
            "const": "audit.matrix.xiak.com/v1"
          },
          "kind": {
            "const": "IngestionResult"
          },
          "outcome": {
            "$ref": "#/components/schemas/IngestionOutcome"
          },
          "record": {
            "$ref": "#/components/schemas/AuditRecord"
          }
        },
        "required": [
          "apiVersion",
          "kind",
          "outcome",
          "record"
        ],
        "type": "object"
      },
      "InstallationVerification": {
        "additionalProperties": false,
        "allOf": [
          {
            "if": {
              "properties": {
                "state": {
                  "const": "PENDING"
                }
              },
              "required": [
                "state"
              ]
            },
            "then": {
              "properties": {
                "eventId": false,
                "fromSequence": false,
                "iamDecisionId": false,
                "recordHash": false,
                "recordSequence": false,
                "toSequence": false
              }
            }
          },
          {
            "if": {
              "properties": {
                "state": {
                  "const": "VERIFIED"
                }
              },
              "required": [
                "state"
              ]
            },
            "then": {
              "required": [
                "eventId",
                "iamDecisionId",
                "recordSequence",
                "fromSequence",
                "toSequence",
                "recordHash"
              ]
            }
          }
        ],
        "properties": {
          "apiVersion": {
            "const": "audit.matrix.xiak.com/v1"
          },
          "checkedAt": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "deploymentId": {
            "type": "string"
          },
          "eventId": {
            "$ref": "#/components/schemas/EventID"
          },
          "fromSequence": {
            "maximum": 9007199254740991,
            "minimum": 1,
            "type": "integer"
          },
          "iamDecisionId": {
            "$ref": "#/components/schemas/DecisionID"
          },
          "installationId": {
            "$ref": "#/components/schemas/ID"
          },
          "kind": {
            "const": "InstallationVerification"
          },
          "operationId": {
            "$ref": "#/components/schemas/OperationID"
          },
          "recordHash": {
            "$ref": "#/components/schemas/Digest"
          },
          "recordSequence": {
            "maximum": 9007199254740991,
            "minimum": 1,
            "type": "integer"
          },
          "state": {
            "$ref": "#/components/schemas/InstallationVerificationState"
          },
          "toSequence": {
            "maximum": 9007199254740991,
            "minimum": 1,
            "type": "integer"
          }
        },
        "required": [
          "apiVersion",
          "checkedAt",
          "deploymentId",
          "installationId",
          "kind",
          "operationId",
          "state"
        ],
        "type": "object"
      },
...
      "QueryRecordsRequest": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "$ref": "#/components/schemas/Action"
          },
          "actor": {
            "$ref": "#/components/schemas/ActorReference"
          },
          "cursor": {
            "$ref": "#/components/schemas/Cursor"
          },
          "from": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "pageSize": {
            "maximum": 200,
            "minimum": 1,
            "type": "integer"
          },
          "to": {
            "$ref": "#/components/schemas/Timestamp"
          }
        },
        "required": [
          "pageSize"
        ],
        "type": "object"
      },
...
      "RecordPage": {
        "additionalProperties": false,
        "oneOf": [
          {
            "properties": {
              "installationId": false
            },
            "required": [
              "tenantId"
            ]
          },
          {
            "properties": {
              "tenantId": false
            },
            "required": [
              "installationId"
            ]
          }
        ],
        "properties": {
          "apiVersion": {
            "const": "audit.matrix.xiak.com/v1"
          },
          "installationId": {
            "$ref": "#/components/schemas/ID"
          },
          "kind": {
            "const": "AuditRecordPage"
          },
          "nextCursor": {
            "$ref": "#/components/schemas/Cursor"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/AuditRecord"
            },
            "maxItems": 200,
            "type": "array"
          },
          "tenantId": {
            "$ref": "#/components/schemas/TenantID"
          }
        },
        "required": [
          "apiVersion",
          "kind",
          "records"
        ],
        "type": "object"
      },
...
      "VerifyChainRequest": {
        "additionalProperties": false,
        "properties": {
          "fromSequence": {
            "maximum": 9007199254740991,
            "minimum": 1,
            "type": "integer"
          },
          "maximumRecords": {
            "maximum": 10000,
            "minimum": 1,
            "type": "integer"
          }
        },
        "required": [
          "fromSequence",
          "maximumRecords"
        ],
        "type": "object"
      },
      "VerifyInstallationRequest": {
        "additionalProperties": false,
        "properties": {
          "deploymentId": {
            "type": "string"
          },
          "installationId": {
            "$ref": "#/components/schemas/ID"
          },
          "operationId": {
            "$ref": "#/components/schemas/OperationID"
          }
        },
        "required": [
          "deploymentId",
          "installationId",
          "operationId"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "InstallationVerifier": {
        "description": "Narrow installation-verifier service credential authorized by IAM installation.verify.",
        "scheme": "bearer",
        "type": "http"
      },
      "ServiceCredential": {
        "description": "Opaque IAM service credential that determines the Audit event source.",
        "scheme": "bearer",
        "type": "http"
      },
      "UserSession": {
        "description": "Opaque IAM user session. Audit derives tenant and subject by calling IAM.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Matrix Audit v1 contracts",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/events": {
      "post": {
        "operationId": "ingestAuditEvent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionResult"
                }
              }
            },
            "description": "Equal event replay."
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionResult"
                }
              }
            },
            "description": "New event accepted."
          },
          "400": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "401": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "409": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "413": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "415": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "422": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "500": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "503": {
            "$ref": "#/components/responses/ProblemResponse"
          }
        },
        "security": [
          {
            "ServiceCredential": []
          }
        ],
        "summary": "Ingest a sanitized event from its authenticated source"
      }
    },
    "/v1/installation:verify": {
      "post": {
        "operationId": "verifyInstallationAudit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyInstallationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallationVerification"
                }
              }
            },
            "description": "Command completed."
          },
          "400": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "401": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "403": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "409": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "413": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "415": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "422": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "500": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "503": {
            "$ref": "#/components/responses/ProblemResponse"
          }
        },
        "security": [
          {
            "InstallationVerifier": []
          }
        ],
        "summary": "Verify the exact fixed PaaS probe Audit fact and chain"
      }
    },
    "/v1/integrity:verify": {
      "post": {
        "operationId": "verifyAuditChain",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyChainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChainVerification"
                }
              }
            },
            "description": "Command completed."
          },
          "400": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "401": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "403": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "409": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "413": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "415": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "422": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "500": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "503": {
            "$ref": "#/components/responses/ProblemResponse"
          }
        },
        "security": [
          {
            "UserSession": []
          }
        ],
        "summary": "Verify a bounded segment of the IAM-derived tenant's chain"
      }
    },
    "/v1/platform/integrity:verify": {
      "post": {
        "operationId": "verifyPlatformAuditChain",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyChainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChainVerification"
                }
              }
            },
            "description": "Command completed."
          },
          "400": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "401": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "403": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "409": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "413": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "415": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "422": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "500": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "503": {
            "$ref": "#/components/responses/ProblemResponse"
          }
        },
        "security": [
          {
            "UserSession": []
          }
        ],
        "summary": "Verify a bounded segment of the IAM-derived installation's chain"
      }
    },
    "/v1/platform/records:query": {
      "post": {
        "operationId": "queryPlatformAuditRecords",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRecordsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordPage"
                }
              }
            },
            "description": "Command completed."
          },
          "400": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "401": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "403": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "409": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "413": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "415": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "422": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "500": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "503": {
            "$ref": "#/components/responses/ProblemResponse"
          }
        },
        "security": [
          {
            "UserSession": []
          }
        ],
        "summary": "Query the IAM-derived installation's Audit records"
      }
    },
    "/v1/records:query": {
      "post": {
        "operationId": "queryAuditRecords",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRecordsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordPage"
                }
              }
            },
            "description": "Command completed."
          },
          "400": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "401": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "403": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "409": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "413": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "415": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "422": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "500": {
            "$ref": "#/components/responses/ProblemResponse"
          },
          "503": {
            "$ref": "#/components/responses/ProblemResponse"
          }
        },
        "security": [
          {
            "UserSession": []
          }
        ],
        "summary": "Query the IAM-derived tenant's Audit records"
      }
    }
  },
  "security": [
    {
      "UserSession": []
    }
  ]
}